requirements.txt
chromadb==0.4.18
python-dotenv==1.0.0
transliterate==1.10.2
sentence-transformers==2.2.2
huggingface_hub==0.16.4
youtube-transcript-api==1.0.3
google-api-python-client==2.166.0